requirements.txt
Flask==2.0.1
Flask-Babel==2.0.0
manim==0.17.3
chainlit
openai
python-dotenv
pdfplumber
pytesseract
pdf2image
